refactor: move handler to retrieve module
